+++ packages/user/src/index.ts
@@ -2,6 +2,13 @@ import "@dzangolab/fastify-mercurius";
 
 import type { SupertokensConfig } from "./supertokens";
 import type { IsEmailOptions, StrongPasswordOptions, User } from "./types";
+import type { ServerResponse } from "node:http";
+
+declare module "fastify" {
+  interface FastifyReply {
+    setHeader: ServerResponse["setHeader"];
+  }
+}
 
 declare module "mercurius" {
   interface MercuriusContext {
@@ -13,6 +20,13 @@ declare module "mercurius" {
 declare module "@dzangolab/fastify-config" {
   interface ApiConfig {
     user: {
+      invitation?: {
+        /**
+         * Number of days in which invitation expire.
+         * @default 30
+         */
+        expireAfterInDays?: number;
+      };
       email?: IsEmailOptions;
       password?: StrongPasswordOptions;
       supertokens: SupertokensConfig;
@@ -32,7 +46,8 @@ export { default } from "./plugin";
 export { default as userResolver } from "./model/users/resolver";
 export { default as UserService } from "./model/users/service";
 export { default as userRoutes } from "./model/users/controller";
-export { default as formatDate } from "./supertokens/utils/formatDate";
+export { default as invitationRoutes } from "./model/invitations/controller";
+export { default as formatDate } from "./lib/formatDate";
 export { default as isRoleExists } from "./supertokens/utils/isRoleExists";
 export { default as validateEmail } from "./validator/email";
 export { default as validatePassword } from "./validator/password";