+++ packages/user/src/model/invitations/handlers/index.ts
@@ -1,0 +1,15 @@
+import acceptInvitation from "./acceptInvitation";
+import createInvitation from "./createInvitation";
+import getInvitationByToken from "./getInvitationByToken";
+import listInvitation from "./listInvitation";
+import resendInvitation from "./resendInvitation";
+import revokeInvitation from "./revokeInvitation";
+
+export default {
+  acceptInvitation,
+  createInvitation,
+  getInvitationByToken,
+  listInvitation,
+  resendInvitation,
+  revokeInvitation,
+};