+++ packages/user/src/supertokens/recipes/config/third-party-email-password/emailPasswordSignIn.ts
@@ -1,5 +1,5 @@
+import formatDate from "../../../../lib/formatDate";
 import UserService from "../../../../model/users/service";
-import formatDate from "../../../utils/formatDate";
 
 import type {
   AuthUser,